feat: Base low health tutorial message
