--- Assets/BoleteHell/Gameplay/Characters/Enemy/AchievementKillFirstEnemy.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 31c6c0d980c2462799e270bd057bc2a4
-timeCreated: 1764867561
