--- Assets/BoleteHell/Utils/Advisor/AdvisorPopup.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: bbc113eb08be4b8e81f37f33d094f55a
-timeCreated: 1764863293
