--- Assets/BoleteHell/Utils/Advisor/ITutorialService.cs
@@ -1,12 +1,0 @@
-﻿namespace BoleteHell.Utils.Advisor
-{
-    public enum TutorialEvent
-    {
-        KilledFirstBolete,
-    }
-    
-    public interface ITutorialService
-    {
-        void ShowTutorial(TutorialEvent tutorialEvent);
-    }
-}
