--- Assets/BoleteHell/Utils/Advisor/ITutorialService.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 127262b687604c44bc48c2445304eba9
-timeCreated: 1764865519
