+++ Assets/Plugins/Zenject/OptionalExtras/ReflectionBaking/Zenject-ReflectionBaking.csproj.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 91b0b4c77fdd1d24aa1af646e9d29960
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
