+++ Assets/Plugins/Zenject/Source/Zenject.csproj.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8645273165cb7f54290d6eaa1e10ab37
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
