+++ Assets/_BoleteAssets/Art/Tutorial/Materials/BaseBase.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d4d3e827714471941b5bdbc45bd6c931
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
