--- Assets/_BoleteAssets/Data/Characters/Default Spawn List.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 687b8d07401848640a989901dc9551eb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
