--- Assets/_BoleteAssets/Data/SpawnLists/SpawnList.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f0cf5c9ebeb306c4d95a227e384a70bf
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
